💄 :: 지각자 Cell
